Update frontend.sh
(1). Provide a prompt for the user to sign-up and sign-in:
	Sign In
		Take Test
		View Test
	Sign up
	Exit

(2). Using sign-up, the user can register with a user-id and password, and we ask for username. The username should contain only alphanumeric symbols and when this condition is not matched the again ask for new password. 

(3). The password can contain any symbol in it. Accept at least eight characters with at least a number and symbol ask the user to re-enter the password again for validation already registered user can sign-in with ID and password.

(4). The script should prompt for username and password, the password should be like a shadow password and show error in-case passwords mismatches.